db 테이블에서 item 조회

+++ src/main/java/com/example/revise/controller/ItemController.java
@@ -22,7 +22,7 @@ public class ItemController {
     }
 
     @GetMapping("/items/{id}")
-    public ItemData.ItemObjRes item(@PathVariable int id) {
+    public ItemData.ItemObjRes item(@PathVariable long id) {
 
         ItemData item1 = itemService.getItem(id);
